backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Mail==0.9.1
Flask-CORS==4.0.0

psycopg2-binary==2.9.9
redis==5.0.1
python-dotenv==1.0.0
itsdangerous==2.1.2
email-validator==2.1.0
gunicorn==21.2.0
alembic==1.13.1
Pillow==10.1.0
python-multipart==0.0.6 